Allow (optionally) forwading all GTM event data to Klaviyo
It might be helpful for some folks to be able to send all the properties
they're setting in their GTM tag onto their Klaviyo instance, to prevent
having to make changes in GTM each and every time they add or change a
property in their application.

You can additionally exclude certain properties from being forwarded,
which is loosely copied from the Google GA4 Event tag. Useful for
avoiding PII being sent to Klaviyo if it made it to the server-container
in GTM.

Closes !3