Don't retrieve and use email from cookie if storing not enabled
Because of the issue fixed in 7140d1a3f2e8d724f2efb732b01fd38dd479557d,
many people will have cookies with this email set. If they then update
to a fixed version and leave this field disabled they may believe email
is *not* being sent to Klaviyo when it may indeed be, even if not
included in event properties.

By only retriving the email from cookie when storing the cookie is
enabled, we sidestep this issue.